backend/requirements.txt
Django
djangorestframework
openai
requests
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
django-cors-headers
django-environ